# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Blog Pessoal",
    "version": "1.0.0",
    "description": "API RESTful completa para um sistema de blog pessoal desenvolvida em Laravel com autenticação JWT como trabalho de conclusão da matéria Sistemas Distribuídos e API da PÓS em FullStack da Unimar.",
    "contact": {
      "name": "Fabiano Souza de Oliveira",
      "email": "fsouzaweb@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Servidor de Desenvolvimento"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Autenticação JWT. Use: Bearer {seu_token_jwt}"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "example": 1},
          "name": {"type": "string", "example": "Fabiano Souza"},
          "email": {"type": "string", "format": "email", "example": "fabiano@blog.com"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "example": 1},
          "name": {"type": "string", "example": "Laravel"},
          "slug": {"type": "string", "example": "laravel"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "Post": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "example": 1},
          "title": {"type": "string", "example": "Meu Primeiro Post"},
          "content": {"type": "string", "example": "Conteúdo do post..."},
          "published_at": {"type": "string", "format": "date-time"},
          "author": {"$ref": "#/components/schemas/User"},
          "category": {"$ref": "#/components/schemas/Category"},
          "tags": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/Tag"}
          },
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "Tag": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "example": 1},
          "name": {"type": "string", "example": "Laravel"},
          "slug": {"type": "string", "example": "laravel"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "Comment": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "example": 1},
          "body": {"type": "string", "example": "Ótimo post, parabéns!"},
          "author": {"$ref": "#/components/schemas/User"},
          "post": {"$ref": "#/components/schemas/Post"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      }
    }
  },
...
  "paths": {
...
    "/categories": {
      "get": {
        "tags": ["Categorias"],
        "summary": "Listar todas as categorias",
        "description": "Retorna uma lista paginada de todas as categorias",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Lista de categorias obtida com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/Category"}
                    }
                  }
                }
              }
            }
          },
          "401": {"description": "Não autenticado"}
        }
      },
      "post": {
        "tags": ["Categorias"],
        "summary": "Criar nova categoria",
        "description": "Cria uma nova categoria no sistema",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {"type": "string", "example": "Laravel 11"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Categoria criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {"$ref": "#/components/schemas/Category"}
                  }
                }
              }
            }
          },
          "422": {"description": "Dados de validação inválidos"}
        }
      }
    },
    "/categories/{id}": {
      "get": {
        "tags": ["Categorias"],
        "summary": "Exibir categoria específica",
        "description": "Retorna os dados de uma categoria específica",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID da categoria",
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "Categoria encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {"$ref": "#/components/schemas/Category"}
                  }
                }
              }
            }
          },
          "404": {"description": "Categoria não encontrada"}
        }
      },
      "put": {
        "tags": ["Categorias"],
        "summary": "Atualizar categoria",
        "description": "Atualiza os dados de uma categoria existente",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID da categoria",
            "schema": {"type": "integer"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {"type": "string", "example": "Laravel 12"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Categoria atualizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {"$ref": "#/components/schemas/Category"}
                  }
                }
              }
            }
          },
          "404": {"description": "Categoria não encontrada"},
          "422": {"description": "Dados de validação inválidos"}
        }
      },
      "delete": {
        "tags": ["Categorias"],
        "summary": "Excluir categoria",
        "description": "Remove uma categoria do sistema",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID da categoria",
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "Categoria excluída com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string", "example": "Categoria excluída com sucesso"}
                  }
                }
              }
            }
          },
          "404": {"description": "Categoria não encontrada"}
        }
      }
    },
...
    "/comments": {
      "get": {
        "tags": ["Comentários"],
        "summary": "Listar comentários",
        "description": "Lista comentários do usuário ou de um post específico",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "post_id",
            "in": "query",
            "description": "ID do post para filtrar comentários",
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de comentários obtida com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/Comment"}
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Comentários"],
        "summary": "Criar novo comentário",
        "description": "Cria um novo comentário em um post (dispara job de notificação)",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["body", "post_id"],
                "properties": {
                  "body": {"type": "string", "example": "Ótimo post, parabéns!"},
                  "post_id": {"type": "integer", "example": 1}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comentário criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {"$ref": "#/components/schemas/Comment"}
                  }
                }
              }
            }
          },
          "422": {"description": "Dados de validação inválidos"}
        }
      }
    },
...
  }
}